% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CovSample.R
\name{CovSample}
\alias{CovSample}
\title{Update the means (MULTIVARIATE)}
\usage{
CovSample(ns_cov, WkZ_cov, ybar_cov, c0, C0)
}
\arguments{
\item{covs}{Covariances}

\item{ns_mu}{input}

\item{ybar_mu}{input}

\item{n0}{input}

\item{b0}{input}
}
\value{
Vector, draws from full conditional distribution of means given covariances,  priors, etc
The values sum to 1.
}
\description{
...
}
\examples{

}
\keyword{Gaussian}
\keyword{multivariate}

