% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dDirichlet.R
\name{dDirichlet}
\alias{dDirichlet}
\title{Density of Dirichlet distribution Function}
\usage{
dDirichlet(x, alpha)
}
\arguments{
\item{x}{Input, must sum to 1}

\item{alpha}{Hyper-parameter}
}
\description{
density of Dirichlet
}
\examples{
dDirichlet(c(.1, .9), c(0.1,0.1))
}
\keyword{dirichlet}

