% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getNk.R
\name{getNK}
\alias{getNK}
\title{Count number of obs per group, including empty}
\usage{
getNK(Z, k)
}
\arguments{
\item{Z}{allocations}
}
\value{
vector
The values sum to n
}
\description{
...
}
\examples{

}
\keyword{count}

