% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSV.R
\name{getSV}
\alias{getSV}
\title{Update / get sv}
\usage{
getSV(Means, y, z)
}
\arguments{
\item{Mu}{Means}

\item{Cov}{input}

\item{Pi}{input}

\item{k}{input}
}
\value{
Matrix of probabilities
}
\description{
...
}
\examples{

}
\keyword{Gaussian}
\keyword{multivariate}

