% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSx.R
\name{getSx}
\alias{getSx}
\title{Get sx parameter}
\usage{
getSx(y, Z, k)
}
\arguments{
\item{y}{Number of groups to include}

\item{Z}{value}

\item{k}{value}
}
\value{
List of values needed to update parameters
The values sum to 1.
}
\description{
...
}
\examples{

}
\keyword{gibbs}
\keyword{initiate}
\keyword{inputs}

