% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getWkZ.R
\name{getWkZ}
\alias{getWkZ}
\title{Within group unexplained variability}
\usage{
getWkZ(y, Z, k)
}
\arguments{
\item{y}{Number of groups to include}

\item{Z}{allocations}

\item{k}{value}

\item{n}{value}

\item{r}{value}
}
\value{
List of values needed to update parameters
The values sum to 1.
}
\description{
...
}
\examples{

}
\keyword{gibbs}
\keyword{initiate}
\keyword{inputs}

