% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getYbar.R
\name{getYbar}
\alias{getYbar}
\title{GetYbar, MVN}
\usage{
getYbar(y, Z, k)
}
\arguments{
\item{Z}{allocations}
}
\value{
List of values needed to update parameters
The values sum to 1.
}
\description{
...
}
\examples{

}
\keyword{gibbs}
\keyword{initiate}
\keyword{inputs}

