% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initiate_univ.R
\name{initiate_univ}
\alias{initiate_univ}
\title{Initiate the values for MCMC (UNIVARIATE)}
\usage{
initiate_univ(y, k, init.method = c("Kmeans", "uniform"))
}
\arguments{
\item{y}{data, as a vector}

\item{k}{Number of groups in model}

\item{init.method}{Method for choosing initial annotations, "Kmeans"  fits a k component Kmeans to the data, while "uniform" spreads k groups evenly across the range of the data.}
}
\value{
List of values needed to start MCMC.
The first element is a tibble containing mu0, sig0, and p0. The second, a vector of integers containing the allocations init.Z.
}
\description{
...
}
\examples{
 y <- rnorm(10,1, 10)
 initiate_univ(y, k=2, init.method="uniform")
 initiate_univ(y, k=5, init.method="uniform")
 initiate_univ(y, k=5, init.method="Kmeans")
}
\keyword{gibbs}
\keyword{initiate}
\keyword{inputs}

