% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initiate_values.R
\name{initiate_values}
\alias{initiate_values}
\title{Initiate the allocations for MCMC}
\usage{
initiate_values(Z, y, k, r)
}
\arguments{
\item{Z}{allocations}

\item{y}{Number of groups to include}

\item{k}{value}

\item{r}{value}

\item{n}{value}
}
\value{
List of values needed to update parameters
The values sum to 1.
}
\description{
...
}
\examples{

}
\keyword{gibbs}
\keyword{initiate}
\keyword{inputs}

