% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loglike_UNIV.R
\name{loglike_UNIV}
\alias{loglike_UNIV}
\title{Update the allocation probabilities for multivariate Gaussian mixture}
\usage{
loglike_UNIV(.Zs, y, Ps, Mus, Cov)
}
\arguments{
\item{Cov}{input}

\item{Mu}{Means}

\item{Pi}{input}

\item{k}{input}
}
\value{
Matrix of probabilities
}
\description{
...
}
\examples{

}
\keyword{Gaussian}
\keyword{multivariate}

