% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parallelAccept.R
\name{parallelAccept}
\alias{parallelAccept}
\title{Update the allocation probabilities for multivariate Gaussian mixture}
\usage{
parallelAccept(w1, w2, a1, a2)
}
\arguments{
\item{Mu}{Means}

\item{Cov}{input}

\item{Pi}{input}

\item{k}{input}
}
\value{
Matrix of probabilities
}
\description{
...
}
\examples{

}
\keyword{Gaussian}
\keyword{multivariate}

