% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdirichlet.R
\name{rdirichlet}
\alias{rdirichlet}
\title{Draw from Dirichlet distribution}
\usage{
rdirichlet(par, m = 1)
}
\arguments{
\item{par}{Vector of inputs (Alpha+ number of obs in each group)}

\item{m=1}{number of vectors to generate}
}
\value{
Vector of probabilities, of same length as \code{par}.
The values sum to 1.
}
\description{
This function draws samples from a dirichlet distriution with hyperpars m.
}
\examples{
rdirichlet(c(1,1,1))
rdirichlet(c(0.5,20,0.000000001))
}
\keyword{dirichlet}

