% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimMVN.R
\name{SimMVN}
\alias{SimMVN}
\title{Simulate a multivariate Gaussian mixture}
\usage{
SimMVN(Means, Covs, P, N)
}
\arguments{
\item{Means}{Means}

\item{Covs}{input}

\item{P}{weights}

\item{N}{input}
}
\value{
list
}
\description{
...
}
\examples{

}
\keyword{Gaussian}
\keyword{multivariate}

