% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Update_probs.R
\name{Update_probs}
\alias{Update_probs}
\title{Update the allocation probabilities for multivariate Gaussian mixture}
\usage{
Update_probs(Mu, Cov, Pi, k)
}
\arguments{
\item{Mu}{Means}

\item{Cov}{input}

\item{Pi}{input}

\item{k}{input}
}
\value{
Matrix of probabilities
}
\description{
...
}
\examples{

}
\keyword{Gaussian}
\keyword{multivariate}

