% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Update_Zs.R
\name{Update_Zs}
\alias{Update_Zs}
\title{Update the Allocations}
\usage{
Update_Zs(PZs)
}
\arguments{
\item{covs}{Covariances}

\item{ns_mu}{input}

\item{ybar_mu}{input}

\item{n0}{input}

\item{b0}{input}
}
\value{
Vector, draws from full conditional distribution of means given covariances,  priors, etc
The values sum to 1.
}
\description{
...
}
\examples{

}
\keyword{Gaussian}
\keyword{multivariate}

