% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initiate_Z.R
\name{initiate_Z}
\alias{initiate_Z}
\title{Initiate the allocations for MCMC}
\usage{
initiate_Z(x, k, n, method = c("Kmeans", "random", "single"))
}
\arguments{
\item{x}{data}

\item{k}{Number of groups to include}

\item{n}{value}

\item{method}{Type of initiallization: Kmeans, random, or single}

\item{m=1}{number of vectors to generate}
}
\value{
Vector of inital allocations for starting mcmc
The values sum to 1.
}
\description{
...
}
\examples{

}
\keyword{dirichlet}

